requirements/common.txt
#Django
django>=4.0,<4.1
djangorestframework>=3.13.0
djangorestframework-camel-case<=1.3.0
psycopg2-binary>=2.8
drf-yasg

#Utils
flake8
black
isort==5.8.0